CEF控件离屏渲染模式：设置页面光标的代码，统一放在平台相关的类中实现。
