+++ duilib/CEFControl/CefWindowUtils_Windows.cpp
@@ -143,6 +143,15 @@ bool CaptureCefWindowBitmap(CefWindowHandle cefWindow, std::vector<uint8_t>& bit
     return true;
 }
 
+void SetCefWindowCursor(CefWindowHandle cefWindow, CefCursorHandle cursor)
+{
+    if ((cefWindow == nullptr) || (cursor == nullptr)) {
+        return;
+    }
+    ::SetClassLongPtr((HWND)cefWindow, GCLP_HCURSOR, static_cast<LONG>(reinterpret_cast<LONG_PTR>(cursor)));
+    ::SetCursor(cursor);
+}
+
 } //namespace ui
 
 #endif //DUILIB_BUILD_FOR_WIN
